fixes#8098
